test: Move test helpers into shared lib
